Refactor GuardDuty to waiter pattern
